Prevent division by 0 and crash of service on too early or bad sensor read
On a live environment may happen that the method "get_temp" fail with a division per zero error on boot, this causes the service to stop.
This could be related to a too early try to read the sensor or something related to loading of services and sensors.
This always happen on a live environment ISO.
This could happen also if sensor is broken or defective. On a zero read the script should put fan on max just to be sure that it keeps everything cool as ErikaFluff perl script do.